+++ src/commands/generateSourcekitConfiguration.ts
@@ -47,7 +47,7 @@ export async function generateSourcekitConfiguration(ctx: WorkspaceContext): Pro
         await Promise.all(
             foldersToGenerate.map(folder => createSourcekitConfiguration(ctx, folder))
         )
-    ).reduceRight((prev, curr) => prev || curr);
+    ).reduceRight((prev, curr) => prev || curr, false);
 }
 
 export const sourcekitFolderPath = (f: FolderContext) => join(f.folder.fsPath, sourcekitDotFolder);